backend/requirements.txt
sqlalchemy[asyncio]>=2.0.28
asyncpg>=0.29.0
alembic>=1.13.1
fastapi>=0.110.0
uvicorn[standard]>=0.27.1
pydantic>=2.6.3
pydantic-settings>=2.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
httpx>=0.27.0
redis>=5.0.3
aiocache>=0.12.2
authlib>=1.3.0
slowapi>=0.1.9
prometheus-fastapi-instrumentator>=6.1.0
structlog>=24.1.0
numpy>=1.26.4
scipy>=1.13.0
